Move table loggers to a package without "iris".
